Modularize spells and tools
